[FE-127] Moved the Repay componenet to the Buttons component.
