+++ src/components/Buttons/Buttons.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import Borrow from './Borrow/Borrow';
 import Deposit from './Deposit/Deposit';
 import Redeem from './Redeem/Redeem';
+import Repay from './Repay/Repay';
 
 function Buttons({ account }) {
 	return (
@@ -9,6 +10,7 @@ function Buttons({ account }) {
 			<Deposit account={account} />
 			<Redeem account={account} />
 			<Borrow account={account} />
+			<Repay account={account} />
 		</div>
 	);
 }
