+++ src/components/Buttons/Repay/Repay.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import RepayAll from './RepayAll/RepayAll';
+import RepayUnderlyingAsset from './RepayUnderlying/RepayUnderlying';
+
+function Repay({ account }) {
+	return (
+		<div>
+			<RepayAll account={account} />
+			<RepayUnderlyingAsset account={account} />
+		</div>
+	);
+}
+
+export default Repay;
